[flow] Scope analysis for `declare global {...}`
Summary:
This diff adds scope analysis for local checking of `declare global {...}`.

Note that we don't have to do it for normal modules, since by the time we are checking normal modules, it will already have the extended globals injected into the context from type sig, so it doesn't have to read the globals from nested `declare global` block.

For `declare global` within `declare module`, we special case it during scope analysis. We first hoist all the globals within declare global within the declare module, create a scope with it, and then create a scope for everything else. This will implement the scoping rule as described in D68246824.

With this setup, we have the feature mostly working, and the only two issue remainings are globalThis interaction, and missing error for having value-decls within the declare global block.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68851619

fbshipit-source-id: 1122ee2ec84fb0b6df6e7a3f47529ce3c2be6e50