[flow] Fix spurious error w/ StringPrefix indexers in object subtyping
Summary:
When there is an indexer in a lower bound and optional properties in the upper bound we do not necessarily need to flow the indexer to the optional properties if their keys are disjoint. This diff introduces a check to see if we should bother flowing the value types if the indexers do not match up.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D69136858

fbshipit-source-id: c67e45dac041f3e4ce13ded3f27fa18b692566eb