[flow] Fix libdef recheck bug that drops new file updates
Summary:
We didn't include updates when doing a full libdef recheck. This is usually fine because all files will be rechecked anyways. However, it won't be fine if the updates introduce new files, then these will be incorrectly dropped.

This is a hard to repro bug for normal users, because they usually cannot change libdef and add new files simutaneously. However, it will be an issue on CI when all updates are given to flow in batch via `force-recheck`.

This diff fixes the issue by including the full updates.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D69216402

fbshipit-source-id: 054f6f50448723d8a9c36f0cf82783a0214fe1db