[flow] Allow non-covered path for calls to namespaces_bitset_of_path
Summary:
In all the places we call `namespaces_bitset_of_path`, it doesn't always make sense to fail when we cannot map the path to a haste namespace.

It only makes sense when we already know that the file is a haste file (the call site `exported_module`).

However, for the other two call sites:

- In `commandHandler.ml > mk_module_system_info > is_package_file` site, there is no guarantee that the path it's testing will be a haste file, and it already returns an optional, so we should tolerate it.
- In `module_js.ml > resolve_haste_module`, we are trying to get the haste_namespace of the importing file. The importing file might not be a haste file, and `resolve_haste_module` might fail resolve to any haste module, because something like look like haste might just be a node package. (e.g. import `react-relay`). Therefore, when we cannot find haste-namespace mapping, it should be ok as well.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D69146724

fbshipit-source-id: 56e198d041563533c6bcd3a1e96f4e4b25937909