[flow] Refactor 'with_bindings' in env
Summary:
Before for `with_bindings` (and `with_scoped_bindings`), we were passing in a function and an argument, and then within `with_bindings` just calling that function with that argument. Now we just pass in a function that is called.

This will make it possible to actually make use of the return value of the passed in function within name_resolver to simplify some code.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D69407430

fbshipit-source-id: 0f6e4eb545028f99839961b2d27c5f87d24f7c96