[flow][match] Provider and refinement analysis for match expression case bodies
Summary:
While unlikely to be ever used this way, match expression body expressions can be an assignment expression (or a `.push`), so we should support all the things that come along with that (see test cases).

While not likely to be useful for match expressions, this is the majority of the work for getting match statements working, as there the case body is a block rather than one expression - once this code is made more generic.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D69407426

fbshipit-source-id: 8ccce013b64c40c622f402c624aeede2107e9b5a