[flow][match] Update reason location for exhaustive checking error
Summary:
For the added test, before we pointed at `T`, however we want to point at the definition location so we know what is the missing case.
The exhaustive checking logic will be revamped in the future, but for now this improves the experience of our v0 adopters, when they adopt `match` statements.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D69504653

fbshipit-source-id: 720aaca00c73f407a010ee6e6b9e17b985cb4ba5