[flow][tests] Consistent no-flowlib behavior for tests
Summary:
Currently for saved state tests, we have an inconsistent `no-flowlib` setup. During the initial state saving, we run it with the default on `--no-flowlib`, but when running the test script, we don't pass the flag (because there was no way to pass it by environement variables). It caused an inconsistency between saved-state and the real state.

Until now, it just happens to work, because we strip most of the libdef stuff from saved-state. But it will break in the next diff when we will add libdef stuff into saved state. This diff adds env-var support for `--no-flowlib`, so that the entire test suite with saved state can run in a consistent environment.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D69599623

fbshipit-source-id: 9445d7f1ba52f139ccda2273965c2c985c88ba20