[flow] Properly normalize and denormalize libdef paths for saved state to avoid removing libdef from dep-graph hack
Summary:
Right now, we are seeing a `Leader_not_found` crash when checking libdef files after saved-state init. This is caused by bad interactions between the following two:

- In 2023, when I moved libdef checking to normal checking, I intentionally excluded libdef from dependency graph because it seems easy back then, because back then libdef indeed doesn't depend on anything.
- In D69146723, libdef will now depend on the react module.

Combining these two, and with a react module from the userland, we end up with fire, because we end up with a dependency graph that's missing edges from the reality (`libdef file -> flowtyped/react.js.flow`). In the step of ensure-parsed, we are supposed to use these edges to parse files that are skipped because it hasn't changed since saved state. When these edges that should exist but doesn't, we will miss these parses, and miss the relevant entry in the shared memory.

To properly fix this, in this diff, I will stop the hack that removes the libdef dependency graph node in the shared memory and later adds it back with `libdef file -> {}`. Instead, we will attack the problem at front. Previously the issue is that builtin libdef paths are unstable across machines if we try to relativize it against the root (e.g. `/tmp/flow/flowlib_<some_hash>` against either `/users/foo` or `/users/bar` or `/home/users/baz`). This diff will detect such builtin libdef files, and normalize it with a special prefix, so that during denormalization time, we will absolutize it against the builtin flowlib path instead of the root.

With this fix, the saved state will represent the true dep graph, so we don't need other hacks downstreams.

Changelog: [fix] Fixed a crash after initialization with saved state if the typing for `react` module is provided as a userland module.

Reviewed By: jbrown215, panagosg7

Differential Revision: D69577613

fbshipit-source-id: bfaa805e8d99be4a905a2dc10620c5dfe8276a3a