[flow] add optional fallback to ExplicitlyIgnored variant
Summary:
This stack will add the capability to provide a backup flowconfig path to use for explicitly ignored files, which will allow us to give IDE support to directory structures like:

```
multiplat/
├── code
│   ├── a.js
│   └── a.native.js
├── .flowconfig // ignores a.native.js
└── other-code
    └──.flowconfig // Checks a.native.js
```

Right now, if we open code/a.native.js we provide no IDE support because it is explicitly ignored by `multiplat/.flowconfig`. This stack will allow us to specify that `a.native.js` is checked by `other-code/.flowconfig` by calling `flow ls --explain`.

NOTE: This is a TEMPORARY solution intended to be a stopgap and will thus not be documented. The real solution here is to unify the flowconfigs and support multiplatform projects properly.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D69859084

fbshipit-source-id: 97df0e2fccaf5d44cfdbefba761d3e123281c8c1