[flow] Include the full set of libdefs in saved state
Summary:
The only reason we are excluding the builtin ones is because its relative path against the root might be unstable across machines. However, the problem has been addressed with the infra added in D69577613, and we are fully capable of normalizing, storing, and denormalizing all libdef paths, so let's reduce the complexity and just store everything.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D69871981

fbshipit-source-id: 2024b362b78bc33ab531e02e2652eab9fe8c89b4