[flow] Only store unordered libs in saved state
Summary:
The only place where it's necessary is we use the old set of libdefs to check whether there are changes in libdef:

https://github.com/facebook/flow/blob/0b8cc5920e4caa93fba1b6436ff17ebbefc49aff/src/services/inference/types_js.ml#L2296

There, we immediately turn an ordered one into an unordered one. Therefore, it's better to just directly store the unordered one.

This refactor will help for scoped libdef required for multiplatform work, where the ordered version might only apply to certain directories (so it will have more complex structures), but the unordered one (usually for the purpose of telling whether a file is a libdef file) will always remain a set.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D70011192

fbshipit-source-id: a4a0d90de0f62ed6fee30a5f7f5ca1fd4dcb17ae