[flow][refactor][EZ] Rename `libs` to `all_unordered_libs`
Summary:
This diff renames the set of all libdef files for clarify.

This refactor will help for scoped libdef required for multiplatform work, where the ordered version might only apply to certain directories (so it will have more complex structures), but the unordered one (usually for the purpose of telling whether a file is a libdef file) will always remain a set.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D70011191

fbshipit-source-id: 4689b41910a38313d8835bc1985d9848e9c77d95