[flow][match] Add error for referencing binding introduced in same pattern
Summary:
Mirroring how the transform actually works, a pattern's bindings are initialized after the entire pattern condition is analyzed. This means a pattern cannot reference a binding introduced in the same pattern.

I believe this refactor will also make it easier to analyze "or" patterns with bindings in the future.

While having both errors per location is not ideal, I think it's OK for now as the user will still be able to solve their problem.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D70025839

fbshipit-source-id: 596a2c34e4410b7775e198f080a99481817cf845