[flow] Remove most of the bundled libdefs
Summary:
Due to the cyclic dependency between dom.js and bom.js, and the fact that react-dom depends on dom.js, I wasn't able to remove majority of dom.js and all of bom.js and cssom.js for now, but I am still able to kill most of the other libdefs and still be self-contained.

Changelog: [lib] Most of the bundled libdefs will no longer be maintained and shipped with Flow. Going forward, they should be downloaded from flow-typed. To maintain the same behavior as before, you should have a `flow-typed.config.json` in the root of your project with the following content:

```
{
  "env": ["node", "dom", "bom", "intl", "cssom", "indexeddb", "serviceworkers", "webassembly"]
}
```

See https://flow-typed.github.io/flow-typed/#/env-definitions

Reviewed By: jbrown215

Differential Revision: D70197722

fbshipit-source-id: 7ffe8d668239ef7350d001715e1e0e6fa5838084