[flow] Rollback ability to create userland react module typing and kill unreleased declare global support
Summary:
Unfortunately, the userland typing for the react module has caused too many reliability issues, so this diff rolls back all the risky part of the changes, including

- declare global support
- making every module depend on react: for places where we do the `Context.find_require` lookup, we now change it to just builtin lookup.

This means we cannot completely ban override. However, it's less of an issue for multiplatform work, since the override we really need to ban is a directory-scoped overriding the builtin + common libdefs. We should still issue errors on these overrides to prevent surprises.

Changelog: [internal] The ability to have userland React module type definition is never announced.

Reviewed By: gkz

Differential Revision: D70270290

fbshipit-source-id: 54212086b31b3f4bd907d0ba8fd2d841a8931e33