[flow] relax recursively_resolvable check
Summary:
In upcoming changes we will be attaching hints to identifiers to help determine whether we need to generalize types or not (see D66659579 for definitions).

This change would cause some new [definition-cycle] and [recursive-definition] errors, due to `recursively_resolvable` considering, since expression when `chain` is `true` as non-resolvable.

I can't think of a reason why this should be the case, so this diff changes `recursively_resolvable` on `ExpressionDef` to descend into expressions even if there is a hint and when `chain` is `true`. The latter is typically the case when we have a `Member` expression. `expression_resolvable` handles the case of `Member`, so this should also be okay.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D69557697

fbshipit-source-id: 3853e611240c6d63a0d6701f0f116a09e1c2b8c5