[flow][refactor] Rename and make the name already bound error more detailed
Summary:
In the next diff, we are going to error on libdef overrides using the same signature binding error. This diff improves the current one to give it a more specific name, and also does some driveby cleaning to make the error more detailed by pointing our the existing binding.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D70570825

fbshipit-source-id: 0a75674e81414ade595836d5c86788e4d3a2f8f0