[flow][name-def] record hints for more expressions
Summary:
This diff tries to answer a question that comes up in natural inference: *when should a (const-like variable initialized with a) primitive literal be generalized?*

Consider for example
```
const x = "a";
x as "a";
```
In this example, we want the use of `x` to *not* be generalized.

What makes this case special is that it interacts with an annotation. Typically, contexts that interact with annotations have hints available to them. Therefore, we can use the presence of a hint as a proxy for whether the singleton type should not be generalized.

To make this happen, we need to record hints for identifiers. This is what this diff does and goes a few steps further, by recording hints for a few more expressions that could contain (const-vars of) primitive literals.

Consider for example the following code
```
const obj: T = {f: 'a'};
```
The expression `{f:'a'}` is considered synthesizable, so the `hints` passed down its subexpressions in `visit_expression` would be `[]`. Under natural inference, we need to know if the type for `'a'` will hit an annotation. So to restore this information we store the hint for the loc of the object literal. Later on (in a follow up) diff, the recursion into various expressions in statement.ml will pass along the information that an enclosing expression of `'a'`, here `{f:'a'}` has a hint, and therefore we need to infer a precise type for it.

An alternative here would be to pass all hints (`hints_before_synthesizable_check`) to properties of object literals, and then compute the hints for `'a'` directly. This might seen a bit simpler, but it did cause a significant perf regression, as we ended up computing too many hints at the lead nodes of large object trees.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D66659580

fbshipit-source-id: 454681a1527ca7a814a36c55f801d32b2b01e0ff