[flow] add "expected_only" to lazy_hint
Summary:
Under natural inference, we only want to use expected hints to inform the precision of primitive literal inference. For example in
```
const VAL = "value" as const;
var x = cond ? VAL: "a";
x = "b";
```
`VAL` passes a non-expected (aka best-effort) hint to `"a"`. We do not want this hint to force `"a"` to be inferred as the singleton type `"a"`. Instead we would like to infer `string` for that portion of `x` so that the assignment in the next line to succeed.

To enable this behavior, we add a flag to `lazy_hint` that only computes expected hints, and returns `DecompositionError` on best-effort ones.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D70675870

fbshipit-source-id: fbc5c8b8a449b85b0c3cca77c4ba32a7389e8ca6