[flow] Scoped libdef support 4/n: Error on libdef name override
Summary:
With the support for scoped libdefs, the problem of libdef override will be much more severe. Previously, it's just bad that we just forcefully replace one of the defined entries. Now with scoped libdefs where common code is checked only under common libdefs, a scoped libdef binding overriding a common libdef binding in an inconsistent way might create a false sense of security.

Therefore, in this diff, we will now error on the overriding location as a first step, so that we can at least know where the override happens. We can relax this error later when we support more forms of declaration merging.

To detect the name overrides, we intercept at the `bind` function. Whenever we see an illegal bind at the global scope level, we will report an error. I have updated all the updaters (those `f`s) to also return whether the binding is legal.

Changelog: [errors] Overriding already defined names in library definition will now error.

Reviewed By: panagosg7

Differential Revision: D70570824

fbshipit-source-id: 62e59e108143c5233367de626a1af98706bdc660