[flow] handle NullT|VoidT ~> TestPropT consistently
Summary:
It was common before to have groups of constraints of the form
```
NullT ~> TestPropT
VoidT ~> TestPropT
```
While both of these lead to errors, only the first one went through the `GetPropT` route. The other one went thought the catch-all case. The result was getting the exact same error but with two different error codes ([incompatible-use] and [incompatible-type]).

This diff makes both constraints be handled in a similar way through the TestPropT constraint.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D70738169

fbshipit-source-id: 3b56aef4d99b2f0a7ff020af7026757d193328b0