[flow] add has_hint to syntactic_flags
Summary:
The presence of a hint will be important in natural inference to decide whether we need to infer a primitive literal in a precise or general manner. The rationale is that if a hint exists for the position of a literal, then we know that the type of literal will be checked against the type where the hint originally came from, and so we'd want to infer a precise type for it to improve our chances of the check succeeding.

To ensure we have enough information in `Statement.identifier` and `Statement.string_literal`, etc. this diff adds a "has_hint" flag to `syntactic_flags` and adds a `loc_has_hint` function in `Primitive_literal` (trivially returning `false` for now).

Why do we need both of these changes?

Literals and identifiers have been considered synthesizable expressions in LTI. Therefore until recently we didn't really store hints for them. In D66659580, we started recording hints for this expressions, so that they can later be queried by  `loc_has_hint`.

So, why do we need `has_hint`? There are two reasons for this:
1. Even after D66659580 not all identifiers/literals are guaranteed to have hints.
2. Calls to `loc_has_hint` can be expensive. If the identifiers/literals are part of a container that *has* a hint, then it is safe to assume that the "has hint" property holds for them as well. This allows us to short-circuit this check and get the result for each identifier/literal much cheaper. In an earlier version of this stack that didn't include this flag, computing all these hints caused a significant perf regression.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D70676103

fbshipit-source-id: 8183265a8ab27219ea2284fbd3bc34b2fe36d3e1