[flow] add a target_loc field to Context.SynthesisMode
Summary:
In upcoming changes in natural inference we will be computing hints more aggressively, including on expressions that used to be considered synthesizable (primitive literals and variables). These hints may end up being be dependent on each other. Consider for example the call to `foo(x => {}, 'a')`. Computing a hint for `'a'` means we have to run a "synthetic" call to `foo` (see `Instantiate_Callee` in type_hint.ml). This call itself will require a (synthetic) type for its arguments, which include `'a'`. We need to be able to detect this cycle and return a solution for this recursive request for a type for `'a'`.

This diff does some of the preparatory work for this. It stores the target location of synthesis in the `SynthesisMode` constructor. Later on we'll use this location to bail-out of hint computation for the `loc_has_hint` query.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D70763511

fbshipit-source-id: 4665086407fbe80dff214dacb8c0c63e3ec8cc75