[flow] Prevent crash in libdef files for with illegal import export
Summary:
Close https://github.com/facebook/flow/issues/9262

In parsing service, we unconditionally call the module version of type sig parsing even on libdef files:

https://github.com/facebook/flow/blob/9d523198696e2560d9423199f91de34a8cabd19f/src/parsing/parsing_service_js.ml#L156-L165

It will cause a mismatch between the import/export info computed from file_sig and type_sig, and results in a crash like the one in the issue and the mini-repro added in the diff.

We can prevent it by always producing an empty type sig.

The issue starts to appear since 0.260.0. I believe [this commit](https://github.com/facebook/flow/commit/ff29c7da2ec987665867b71b9fdbdb56113a792e) is the one that breaks the parity, and this diff should be able to restore it in a more principle way than the previous parity that happens to work.

Changelog: [fix] Fix a potential crash in libdef files with illegal import export. Thanks techieshark for the repro.

Reviewed By: panagosg7

Differential Revision: D71039980

fbshipit-source-id: 1186050e54c86fe7092c5b8a5f2349532f5d9042