[flow] do not consider annotation singletons as parts of enums
Summary:
Under natural inference, singleton types (e.g. `SingletonStrT`) will be used for both annotations and literal values. There is a `from_annot` field to distinguish the two. The change in this diff makes it so that only annotations are considered for optimized enum construction. This effectively is the same behavior as prior to natural inference.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D71161515

fbshipit-source-id: c38855d201f78e17e835628435d664319da26f19