[flow][match] BigInt properties: parsing + error in object pattern, support in member pattern
Summary:
Flow does not support bigint property names, parse and error on those cases.

For member patterns, we can support bigint, but because of the above statement it's not that useful.

This is already support in Hermes parser.

Also add some `// ERROR` comments I forgot to add in recent diffs.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D71170011

fbshipit-source-id: c27925605b2d6da1a9bc7c1d430ed78beb5ebb10