[flow] keep AST hints in Loc_env
Summary:
Loc_env.hint_map contains lazy hints in the form of `Type.t`s. Using these hints for natural inference can be pretty expensive, so instead we also store the AST version of those hints. Natural inference only needs will only query for the presence of a hint instead of the exact type that it evaluates to.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D71363663

fbshipit-source-id: 294c6b3f692763e8bb60085ecf109ccb25ddcdf3