[flow] natural inference for primitive literals
Summary:
This part of the Natural Inference project is about types we infer for primitive literals and const-variables initialized with primitive literals.

For example, what should the inferred type of the string literals be in the following statements?
```
const x = "a";
const obj = { f: x };
foo("b");
```
The only meaningful options for `"a"` are the general type `string` and the precise singleton type `"a"`. However, picking between the two can be tricky and varies depending on the context and the use of the expression.

Note that the type `StrT_UNSOUND` that is inferred today in `Statement.string_literal_value` combines (in an unsound way) characteristics of both types at the same time.

## The approach

The default behavior under natural inference is to infer the general version of the primitive type. For example, in the definition of `obj` above, `x` will be considered a `string` yielding the type `{f:string}` for `obj`.

However, under certain contexts and conditions the more precise type is needed (`Primitive_literal.needs_precise_type`):

1/ The **enclosing context** is such that requires increased precision: (i) conditional contexts, since this means that some refinement will be performed, and so precise types can give us more accurate results, and (ii) index contexts, since this allows for more precise operations over the object or array that is being accessed. For example,
```
const n = x === "a" ? 1 : 2;
const m = obj[x];
```

2/ We are initializing a **const-variable**. For example
```
const x = 1;
const y = x; // infer y: 1
```

3/ We are in an **`as const`** context. For example
```
const x = 1;
const o = {f: x} as const; // infer {+f:1}
```

4/ We are computing a property of a **frozen object**. Similar to the previous case.

5/ The value is about to be compared against an **annotation**. We use the presence of a hint as a proxy to determine this. The call to `foo("b")` is an example of this. Some special care is needed when `foo` is generic so that the precise type of `"b"` is not leaked to the result. See below.

## Implementation

Diffs leading up to this one have ensure we have the necessary information about context and hints to make the above determinations. Using this info we can call into functions in `Primitive_literals` which adjust the precision of inference. This needs to happen in the following cases:
* Primitive literals (`StringLiteral`, etc.). We call `adjust_precision` directly.
* `Identifier`. This is to handle the case when we need to generalize (`try_generalize`) the type of a const-variable that is initialized with a primitive literal. See example of `obj` above.
* `Conditional`. Precise types that are needed to evaluate refinements in sub-expressions might leak, so we need to generalize appropriately.

In **generic calls** to ensure precise types do not leak to the result, we visit the result of implicit instantiation and replace non-annotation singleton types with their general version. This is what allows us in
```
 declare function useState<S>(init: S): [S, (S)=>void];
const [st, setSt] = useState(42);
```
to infer `st` and `setSt` as `number` and `(number)=>void` instead of the impractical `42` and `(42)=>void`.

Some more notes on implementation:

> Why do we need a "has_hint" field in `syntactic_fields` if we can already call `loc_has_hint`?

If the primitive literal is part of another container literal expression, we might not have a hint recorded for it (since in general recording hints can be costly). For this reason, we first check if the enclosing expression has a hint. This is done by forcing the value of `has_hint` that has been accumulated until this point. For example, in `{f:['a']}`, if the outer object has a hint then `'a'` is also considered to have a hint.

> Why do we still infer `*_UNSOUND` types?

In "partial" mode, the result of generalization is still the previous `*_UNSOUND` types. An exception to this is when generalizing the type of identifiers. See calls to `try_generalize`. "Full" mode will replace all these unsound types with the respective `*GeneralT` version.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D66659598

fbshipit-source-id: 1acd9c41f8a4c6783753a30af3c095647e50d294