Rename `effect` to `effect_` as it is a reserver keyword in OCaml 5.3 (#9268)
Summary:
Fixes https://github.com/facebook/flow/issues/9267

I don't know if any necessary adjustments need to be made to the documentation. We're changing the names of a few types and some labeled arguments, do the docs need to be updated anywhere?

Also, my editor runs `ocamlformat` when I save a file, so the diff might be larger because of that. Let me know if the formatting changes are unwanted, and I can remove them from the commit.

Pull Request resolved: https://github.com/facebook/flow/pull/9268

Reviewed By: panagosg7

Differential Revision: D71492904

Pulled By: SamChou19815

fbshipit-source-id: 117e80c52194040b7428c0726215c725334f384e