[flow] no slow union_compare in union_optimization_guard on synthetic unions
Summary:
`union_compare` in `union_optimization_guard` is a potentially very expensive operation (O(n^2)) as it flattens the input unions first and performs pair-wise checks. When we hit this codepath in scenarios of checks of the form
```
LargeUnion1 | LargeUnion2 <: LargeUnion1 | LargeUnion2
```
then we risk to flatten out the large unions and miss out on the opportunity to check these faster by match `LargeUnion1` on the left with that on the right, and the same for `LargeUnion2`.

In LTI these toplevel unions often emerge in a synthetic when merging values from different branches, as solutions to implicit instantiation problems, and so on. In other words, they do not come from annotations.

This diff tracks synthetic unions (at least most of them) and does not perform the `union_compare` check on them. Instead, it allows the unfolding of the union, which might enable some other optimized path.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D71324568

fbshipit-source-id: af06736206767586f2e66351075d806b44365e34