[flow] do not cache synthesis primitive literal results
Summary:
Similar to how we disable caching when we create placeholders during synthesis mode, we want to disable caching when we infer a general type for primitive literals under synthesis.

This diff does this and also renames the property `synthesis_produced_placeholders` that was used for this purpose before to a more general name (`synthesis_produced_uncacheable_result`) to include the new use.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D71443011

fbshipit-source-id: 3e8fad8b868285870839db92b3e6ff2768e22f43